requirements.txt
bcrypt==3.1.4
chardet==3.0.4
Flask==1.0.2
Flask-Bcrypt==0.7.1
google-api-core==1.9.0
google-auth==1.5.1
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gunicorn==19.9.0
Jinja2==2.10
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.5
requests==2.20.0
rsa==4.0
urllib3==1.24
